feat: book context to jotai
