+++ apps/web/src/routes/kiwi/-header/sidebar/panels/toc.tsx
@@ -1,10 +1,9 @@
 import { ChevronDown, ChevronRight, BookOpen } from "lucide-react";
 import { useState, memo, useCallback } from "react";
 
-import { currentSectionAtom, useAtomValue } from "@bookiwi/jotai";
+import { bookAtom, currentSectionAtom, useAtomValue } from "@bookiwi/jotai";
 
 import { cn } from "#/lib/utils";
-import { useBook } from "#/routes/kiwi/-reader";
 
 interface NavItem {
   href: string;
@@ -97,7 +96,7 @@ function TocItemComponent({
 const MemoizedTocItemComponent = memo(TocItemComponent);
 
 function TocPanel() {
-  const { book } = useBook();
+  const book = useAtomValue(bookAtom);
   const [toc, setToc] = useState<NavItem[]>([]);
   const currentSection = useAtomValue(currentSectionAtom);
   const currentSectionHref = currentSection?.href;
