+++ apps/web/src/routes/kiwi/-reader/contexts/index.ts
@@ -1,3 +1,1 @@
-export * from "./book-context";
 export * from "./record-context";
-export * from "./settings-context";
