+++ apps/web/src/routes/kiwi/-reader/flip-page-button.tsx
@@ -1,9 +1,9 @@
 import { ComponentProps } from "react";
 
-import { useBook } from "./contexts";
+import { bookAtom, useAtomValue } from "@bookiwi/jotai";
 
 function ReaderPrevPageButton(props: ComponentProps<"button">) {
-  const { book } = useBook();
+  const book = useAtomValue(bookAtom);
   const { children, ...rest } = props;
 
   const goToPrevPage = () => {
@@ -26,7 +26,7 @@ function ReaderPrevPageButton(props: ComponentProps<"button">) {
 }
 
 function ReaderNextPageButton(props: ComponentProps<"button">) {
-  const { book } = useBook();
+  const book = useAtomValue(bookAtom);
   const { children, ...rest } = props;
 
   const goToNextPage = () => {
